requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
fpdf2==2.7.9
python-holidays==0.46
python-dotenv==1.0.1
gunicorn==22.0.0
